feat: options component that combines filters + col toggles
